:recycle: Update Nipype version check and automate future updates of said check
