+++ CPAC/utils/versioning/__init__.py
@@ -14,62 +14,7 @@
 
 # You should have received a copy of the GNU Lesser General Public
 # License along with C-PAC. If not, see <https://www.gnu.org/licenses/>.
-"""Gather and report dependency versions alphabetically"""
-from subprocess import PIPE, Popen
-import sys
-from pkg_resources import working_set
-
-
-def cli_version(command, dependency=None, in_result=True, delimiter=' ',
-                formatting=None):
-    """Collect a version from a CLI
-
-    Parameters
-    ----------
-    command : str
-
-    dependency : str, optional
-        software name to report if not included in result
-
-    in_result : boolean, optional
-        parse software name from result?
-
-    delimiter : str, optional
-        if parsing software name from result, what's the delimiter?
-
-    formatting : func, optional
-        if result needs any formatting, function to do so
-
-    Returns
-    -------
-    dict
-        {software: version}
-    """
-    with Popen(command, stdout=PIPE, shell=True) as _command:
-        _version = _command.stdout.read().decode('utf-8')
-    if formatting is not None:
-        _version = formatting(_version)
-    if in_result:
-        return dict([tuple(_version.split(delimiter, 1))])
-    return {dependency: _version}
-
-
-def _version_sort(_version_item):
-    """Key to report by case-insensitive dependecy name"""
-    return _version_item[0].lower()
-
-def first_line(stdout):
-    """Return first line of stdout"""
-    if '\n' in stdout:
-        return stdout.split('\n', 1)[0]
-    return stdout
-
-
-PYTHON_PACKAGES = dict(sorted({
-  d.key: d.version for d in list(working_set)}.items(), key=_version_sort))
-REPORTED = dict(sorted({
-  **cli_version('ldd --version', formatting=first_line),
-  'Python': sys.version.replace('\n', ' ').replace('  ', ' ')
-}.items(), key=_version_sort))
-
-__all__ = ['PYTHON_PACKAGES', 'REPORTED']
+"""Gather and report versions"""
+from CPAC.utils.versioning.dependencies import PYTHON_PACKAGES, REPORTED, \
+                                               REQUIREMENTS
+__all__ = ['PYTHON_PACKAGES', 'REPORTED', 'REQUIREMENTS']
